Adds Create Order to Workflow in first iteration of refactoring for Clean Arc.

+++ CraftBeer/Shared/IntegrationEventsIncoming/PaymentResultEvent.cs
@@ -2,6 +2,6 @@
 {
     public record PaymentResultEvent : IntegrationEventIncoming
     {
-        public decimal Amount { get; init; }
+        public double Amount { get; init; }
     }
 }
